将 prometheus 插件更名为 save 并更新索引
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
